requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-froala-editor==4.0.19
frozenlist==1.3.3
geoip2==4.7.0
h3==3.7.6
idna==3.4
maxminddb==2.3.0
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
timezonefinder==6.2.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
